project/requirements.txt
Flask
Flask-JWT-Extended
psycopg2-binary
Werkzeug
python-dotenv
